TMDM-13613 Investigate the failure performance tests of 13_UpdateExistingRecord_Party

+++ org.talend.mdm.core.storage.sql/src/main/java/com/amalto/core/storage/hibernate/ScatteredTypeMapping.java
@@ -18,6 +18,7 @@ import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 
 import com.amalto.core.storage.record.StorageConstants;
 import org.apache.commons.lang.NotImplementedException;
@@ -75,9 +76,21 @@ class ScatteredTypeMapping extends TypeMapping {
                     Wrapper existingValue = (Wrapper) to.get(referenceFieldMetadata.getName());
                     if (referencedObject != null) {
                         if (existingValue != null) {
-                            session.delete(existingValue);
-                            Wrapper newValue = createObject(contextClassLoader, referencedObject);
-                            to.set(referenceFieldMetadata.getName(), _setValues(session, referencedObject, newValue));
+                            // Check for type change
+                            ComplexTypeMetadata existingType = classLoader.getTypeFromClass(existingValue.getClass());
+                            TypeMapping mapping = mappings.getMappingFromDatabase(existingType);
+                            if (mapping == null) {
+                                throw new IllegalStateException("Type '" + existingType.getName() + "' has no mapping."); //$NON-NLS-1$ //$NON-NLS-2$
+                            }
+                            boolean isSameType = mapping.getUser().equals(referencedObject.getType());
+                            if (isSameType) {
+                                session.evict(existingValue);
+                                to.set(referenceFieldMetadata.getName(), _setValues(session, referencedObject, existingValue));
+                            } else {
+                                session.delete(existingValue);
+                                Wrapper newValue = createObject(contextClassLoader, referencedObject);
+                                to.set(referenceFieldMetadata.getName(), _setValues(session, referencedObject, newValue));
+                            }
                         } else {
                             Wrapper object = createObject(contextClassLoader, referencedObject);
                             to.set(referenceFieldMetadata.getName(), _setValues(session, referencedObject, object));
